Testing plugin version
Changed computers and had to re-download Unity, so I figured it would be a good time to upgrade versions. Had to update the GitHub plugin to make it work. Plugin refused to work until I manually commited with git bash. Testing if it now works properly.

Change itself: small but good practice.